requirements.txt
bumpy>=0.4.3
deliciousapi
flask
flask-markdown
pyscss
pyyaml
